+++ tests/support/payload_builder.py
@@ -3,6 +3,7 @@ class PayloadBuilder:
         self._name = source_name
         self._columns = set()
         self._aggregates = set()
+        self._timeframes = []
         self._joins = {}
         self._column_filters = []
         self._aggregate_filters = []
@@ -31,7 +32,15 @@ class PayloadBuilder:
 
         return self
 
-    def column_filter(self, source_name, name, expression, value):
+    def timeframes(self, *timeframes, join=None):
+        if join:
+            self.join(join).timeframes(*timeframes)
+        else:
+            self._timeframes.extend(timeframes)
+
+        return self
+
+    def legacy_column_filter(self, source_name, name, expression, value):
         filter = {
             "source_name": source_name,
             "name": name,
@@ -42,7 +51,13 @@ class PayloadBuilder:
 
         return self
 
-    def aggregate_filter(self, source_name, name, expression, value):
+    def column_filter(self, key, expression, value):
+        filter = {"key": key, "expression": expression, "value": value}
+        self._column_filters.append(filter)
+
+        return self
+
+    def legacy_aggregate_filter(self, source_name, name, expression, value):
         filter = {
             "source_name": source_name,
             "name": name,
@@ -53,12 +68,24 @@ class PayloadBuilder:
 
         return self
 
-    def order_by(self, source_name, attribute_name, direction):
+    def aggregate_filter(self, key, expression, value):
+        filter = {"key": key, "expression": expression, "value": value}
+        self._aggregate_filters.append(filter)
+
+        return self
+
+    def legacy_order_by(self, source_name, attribute_name, direction):
         order_by_clause = {
             "source_name": source_name,
             "attribute_name": attribute_name,
             "direction": direction,
         }
+        self._order_by_clauses.append(order_by_clause)
+
+        return self
+
+    def order_by(self, key, direction):
+        order_by_clause = {"key": key, "direction": direction}
         self._order_by_clauses.append(order_by_clause)
 
         return self
@@ -77,7 +104,7 @@ class PayloadBuilder:
             "name": self._name,
             "columns": list(self._columns),
             "aggregates": list(self._aggregates),
-            "timeframes": [],
+            "timeframes": self._timeframes,
             "joins": [join.as_join() for join in self._joins.values()],
             "order": self._order_by_clauses,
             "limit": "50",